+++ src/Admin/MetaBox.php
@@ -11,6 +11,10 @@ declare(strict_types=1);
 
 namespace IlloDev\MarkdownNegotiation\Admin;
 
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
 use WP_Post;
 
 /**